refactor: move types to lib dir
